backend/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
celery>=5.3.6
Django==5.1.1
django-cors-headers==4.4.0
django-redis>=5.4.0
djangorestframework==3.15.2
faker==30.8.0
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
joblib==1.4.2
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.23.5
opt_einsum==3.4.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
protobuf==4.25.5
psycopg2-binary==2.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis>=5.0.1
requests==2.32.3
rich==13.8.1
scikit-learn==1.5.2
scipy==1.10.1
six==1.16.0
sqlparse==0.5.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
